Fix ymax and rename Tensorflow nama everywhere
